print/requirements.txt
Pillow>=8.2.0
argparse==1.1
PyBluez>=0.23
qrcode>=6.1
imageio>=2.4.1